// el_aventurero_front/src/shared/interfaces/permissions.enum.ts
export enum AppPermissions {
  ADMIN_ACCESS = 'ADMIN_ACCESS',

  DASHBOARD_ACCESS = 'DASHBOARD_ACCESS',
  MESAS_ACCESS = 'MESAS_ACCESS',
  SALES_ACCESS = 'SALES_ACCESS',

  PRODUCTS_ACCESS = 'PRODUCTS_ACCESS',
  INVENTORY_VIEW = 'INVENTORY_VIEW',
  INVENTORY_COUNT_ACCESS = 'INVENTORY_COUNT_ACCESS',

  SHOPPING_ACCESS = 'SHOPPING_ACCESS',

  CAJA_ACCESS = 'CAJA_ACCESS',
  FLUJO_CAJA_ACCESS = 'FLUJO_CAJA_ACCESS',
  ARQUEO_CAJA_ACCESS = 'ARQUEO_CAJA_ACCESS',

  USERS_ACCESS = 'USERS_ACCESS',
}